[Delivers #50939935] Fix quotes in integration tests
Apostrophes et al. in environments handled by Tralics ('\[ \]', etc.)
get converted to curly apostrophes, whereas apostrophes in math
environments passsed through the pipeline as literal strings
don't get converted. This is OK, as MathJax handles the two cases
in a unified way, and the final appearance is the same regardless
of which type of quote gets used.