Add a second label test
I added a second label test to see if we support putting \label after
the caption instead of inside it:

```latex
\begin{figure}
\includegraphics{bar.png}
\caption{This is another caption.\label{fig:bar}}
\end{figure}
```

versus

```latex
\begin{figure}
\includegraphics{bar.png}
\caption{This is another caption.}
\label{fig:bar}
\end{figure}
```

Both are valid LaTeX, but it turns out that the Tralics parser only
supports the first syntax. This is fine, as it's my preferred
syntax as well (I didn't even realize until a few days ago that
the second one was valid).

Because only one syntax is supported, we don't really need the extra
code in this commit, but it does little harm, so I'm putting it in.

[Delivers #51276617]