Use amsthm for the sake of the PDF
We need to define the codelisting environment in polytexnic-core
for the sake of Tralics, but the default LaTeX theorem environment
is italicized, which isn't what we want for code. The solution
is to use the amsthm from the American Mathematical Society. This
doesn't affect the HTML, but it arranges for the PDF (generated
by `poly build:pdf`) to be unitalicized.

[Delivers #51333043]